subroblock-app/requirements.txt
asgiref==3.3.2
Django==3.2.25
django-rest-swagger
django-filter
drf-yasg
boto3==1.34.44
gevent
gunicorn==22.0.0
gunicorn[gevent]
pytz==2021.1
psycopg2-binary==2.9.9
python_dotenv==1.0.1
sqlparse==0.5.0
whitenoise==5.2.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-ses
cryptography
django-debug-toolbar==3.3.0
